+++ app/db_gui/view/db_gui_menu_bar.rb
@@ -6,6 +6,14 @@ class DbGui
       option :db
   
       body {
+        menu('Query') {
+          menu_item('Run') {
+            on_clicked do
+              db.run_db_command
+            end
+          }
+        }
+        
         menu('Edit') {
           menu_item('Copy Table') {
             enabled <= [db, :db_command_result_rows, computed_by: :db_command_result, on_read: -> (data) { !data.empty? }]
@@ -15,11 +23,11 @@ class DbGui
             end
           }
 
-          menu_item('Copy Table (without headers)') {
+          menu_item('Copy Table (with headers)') {
             enabled <= [db, :db_command_result_rows, computed_by: :db_command_result, on_read: -> (data) { !data.empty? }]
             
             on_clicked do
-              db.copy_table_without_headers
+              db.copy_table_with_headers
             end
           }
 
@@ -31,11 +39,11 @@ class DbGui
             end
           }
 
-          menu_item('Copy Selected Row (without headers)') {
+          menu_item('Copy Selected Row (with headers)') {
             enabled <= [db, :db_command_result_rows, computed_by: :db_command_result, on_read: -> (data) { !data.empty? }]
             
             on_clicked do
-              db.copy_selected_row_without_headers
+              db.copy_selected_row_with_headers
             end
           }